+++ lua/plugins.lua
@@ -43,6 +43,8 @@ return require('packer').startup(function(use)
 
   use 'nvim-treesitter/nvim-treesitter'
 
+  use {'nvim-treesitter/nvim-treesitter-textobjects'}
+
   use {"williamboman/mason.nvim"}
 
   use {"williamboman/mason-lspconfig.nvim"}