+++ changelog.md
@@ -1,5 +1,6 @@
 ### 2022-11-15 二 10:30:17 
 - add monokai theme
+- add nvim-treesitter-textobjects plugin
 
 
 ### 2022-11-14 一 23:39:15